Updated nuget packages, lots more - chaosTool can now add warps - when you double click the tree view nodes, they will auto populate their relevant fields - user has an animation history that takes a tuple^3 as index - all animations are to be sent through client.sendanimation - effects have animations and delays - maps have effects(to be shown on tiles) - new thread in game will handle all persistent effects that will occur on players or map - warps now have a persistent effect animation - the code for warping players based on tile was already in the server, now that warps can be added it is being put to use - chaoslauncher no longer injects dawnd - chaoslauncher no longer references chaos - added a couple warps to the tutorial area for testing (copy dump.rdb into your folder)

+++ MapTool/MapsCache.cs
@@ -178,11 +178,11 @@ namespace ChaosTool
                     writer.Write((ushort)map.Warps.Count);
                     foreach (Chaos.Warp warp in map.Warps.Values)
                     {
-                        writer.Write(warp.SourceX);
-                        writer.Write(warp.SourceY);
+                        writer.Write((byte)warp.SourceX);
+                        writer.Write((byte)warp.SourceY);
                         writer.Write(warp.TargetMapId);
-                        writer.Write(warp.TargetX);
-                        writer.Write(warp.TargetY);
+                        writer.Write((byte)warp.TargetX);
+                        writer.Write((byte)warp.TargetY);
                     }
 
                     //write worldmaps for this map
